html/semantics/forms/attributes-common-to-form-controls/autofocus.html

<!DOCTYPE html>
<html><head>
        <title>autofocus on input element</title>
        <meta content="text/html; charset=UTF-8" http-equiv="content-type">
        <meta content="autofocus on input element" name="description">
    <link href="http://www.w3.org/TR/html5/attributes-common-to-form-controls.html#dom-fe-autofocus" rel="help">
</head>
    <body>
	<script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <div id="log"></div>

    <input autofocus="" style="display: none" type="text">
    <input autofocus="false" style="display: none" type="text">
    <input style="display: none" type="text">
    
    <script type="text/javascript">
        test(function() {assert_true(document.getElementsByTagName("input")[0].autofocus)}, "autofocus on input element on input element");
        test(function() {assert_true(document.getElementsByTagName("input")[1].autofocus)}, "autofocus returns false on input element");
        test(function() {assert_equals(document.getElementsByTagName("input")[1].getAttribute("autofocus"), "false")}, "autofocus returns false when retrieved using getAttribute");
        
        test(function() {assert_false(document.getElementsByTagName("input")[2].autofocus)}, "input element without autofocus attribute returns false");
        test(function() {assert_equals(document.getElementsByTagName("input")[2].getAttribute("autofocus"), null)}, "input element's autofocus attribute is null when not specified");
        
        document.getElementsByTagName("input")[2].autofocus = true;
        
        test(function() {assert_true(document.getElementsByTagName("input")[2].autofocus)}, "autofocus on input element is true when set via script");
        test(function() {assert_equals(document.getElementsByTagName("input")[2].getAttribute("autofocus"), "")}, "accessing an input element's autofocus attribute is an empty string");
        
    </script>
    

</body></html>
